+++ crates/cli/tests/activate_test.rs
@@ -100,7 +100,7 @@ bun = "1.1.0"
 
         let output = get_activate_output(&assert, &sandbox);
 
-        assert!(output.contains("<WORKSPACE>/.proto/activate-start:<WORKSPACE>/.proto/shims:<WORKSPACE>/.proto/bin:<WORKSPACE>/.proto/activate-stop"));
+        assert!(output.contains("<WORKSPACE>/.proto/activate-start <WORKSPACE>/.proto/shims <WORKSPACE>/.proto/bin <WORKSPACE>/.proto/activate-stop"));
     }
 
     #[test]
@@ -184,7 +184,7 @@ KEY2 = "value2"
 
             let output = get_activate_output(&assert, &sandbox);
 
-            assert!(output.contains("<WORKSPACE>/.proto/activate-start:<WORKSPACE>/.proto/shims:<WORKSPACE>/.proto/bin:<WORKSPACE>/.proto/activate-stop"));
+            assert!(output.contains("<WORKSPACE>/.proto/activate-start <WORKSPACE>/.proto/shims <WORKSPACE>/.proto/bin <WORKSPACE>/.proto/activate-stop"));
         }
     }
 }